+++ docs/source/releasenotes.rst
@@ -9,6 +9,12 @@ Release notes
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+20.1.1 - 25 Jan 2023
+--------------------
+
+- Library **RPA.Browser.Selenium** (:issue:`781`): Ability to determine the
+  default search order of browsers, when using the ``Open Available Browser`` keyword.
+
 20.1.0 - 19 Jan 2023
 --------------------